requirements.txt
openai
langchain
langchain-openai
langchain-groq
langchain-community
streamlit
python-dotenv
PyPDF2


-e .